feat(controls-island): add useControlsIslandCollapse composable and related tests
