+++ packages/stage-ui/src/composables/index.ts
@@ -9,6 +9,7 @@ export * from './use-breakpoints'
 export * from './use-build-info'
 export * from './use-chat-session/summary'
 export * from './use-controls-island-auto-hide'
+export * from './use-controls-island-collapse'
 export * from './use-inference-preload'
 export * from './use-inference-status'
 export * from './use-lamp-flicker-animation'
